+++ packages/cli/src/rename/data-set/DataSet.handler.ts
@@ -1,0 +1,31 @@
+/**
+ * This program and the accompanying materials are made available under the terms of the
+ * Eclipse Public License v2.0 which accompanies this distribution, and is available at
+ * https://www.eclipse.org/legal/epl-v20.html
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Copyright Contributors to the Zowe Project.
+ *
+ */
+
+import type { IHandlerParameters } from "@zowe/imperative";
+import type { ds, ZSshClient } from "zowe-native-proto-sdk";
+import { SshBaseHandler } from "../../SshBaseHandler";
+
+export default class RenameDataSetHandler extends SshBaseHandler {
+    public async processWithClient(params: IHandlerParameters, client: ZSshClient): Promise<ds.RenameDatasetResponse> {
+        const dsnameBefore = params.arguments.dsnameBefore;
+        const dsnameAfter = params.arguments.dsnameAfter;
+
+        const response = await client.ds.renameDataset({ dsnameBefore, dsnameAfter });
+
+        const dsMessage = `Dataset "${dsnameBefore}" renamed to "${dsnameAfter}"`;
+        params.response.data.setMessage(dsMessage);
+        params.response.data.setObj(response);
+        if (response.success) {
+            params.response.console.log(dsMessage);
+        }
+        return response;
+    }
+}